added 20 CC Musicbox 3.0

+++ src/at/stl/Musicbox/Title.java
@@ -2,9 +2,9 @@ package at.stl.Musicbox;
 
 public class Title {
     public String title;
-    private float length;
+    private double length;
 
-    public Title(String title,float length){
+    public Title(String title,double length){
         this.title = title;
         this.length = length;
     }
@@ -13,6 +13,8 @@ public class Title {
     public void setTitle(String title) {this.title = title;}
     public String getTitle() {return title;}
 
-    public void setLength(float length) {this.length = length;}
-    public float getLength() {return length;}
+    public void setLength(double length) {this.length = length;}
+    public double getLength() {return length;}
+
+
 }
